DEV: Move custom post message callbacks into its own initializer.
